requirements.txt
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.41.1
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.28.2
PyMatting==1.1.12
PyWavelets==1.4.1
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.21.0
scipy==1.10.1
sympy==1.13.3
tifffile==2023.7.10
tqdm==4.66.5
urllib3==2.2.3
zipp==3.20.2
